moved /openapi to ServiceInfoController
