+++ src/test/groovy/io/seqera/wave/controller/ServiceInfoControllerTest.groovy
@@ -18,24 +18,27 @@
 
 package io.seqera.wave.controller
 
-import groovy.transform.CompileStatic
+import spock.lang.Specification
+
 import io.micronaut.http.HttpResponse
-import io.micronaut.http.annotation.Controller
-import io.micronaut.http.annotation.Get
-import io.micronaut.scheduling.TaskExecutors
-import io.micronaut.scheduling.annotation.ExecuteOn
+import io.micronaut.http.HttpStatus
+
 /**
- * Implements a controller for openapi
  *
  * @author Munish Chouhan <munish.chouhan@seqera.io>
  */
-@CompileStatic
-@Controller
-@ExecuteOn(TaskExecutors.BLOCKING)
-class OpenAPIController {
-    @Get(uri = "/openapi")
-    HttpResponse getOpenAPI() {
-    println("Redirecting to /openapi/")
-        HttpResponse.redirect(URI.create("/openapi/"))
+class ServiceInfoControllerTest extends Specification {
+
+    def 'should redirect to /openapi/'() {
+        given:
+        def controller = new ServiceInfoController()
+
+        when:
+        HttpResponse response = controller.getOpenAPI()
+
+        then:
+        response.status == HttpStatus.MOVED_PERMANENTLY
+        response.header('Location') == '/openapi/'
     }
+
 }
